+++ Falcon9_example.py
@@ -1,34 +1,75 @@
-# Date: 06/01/2019
+# Date: 16/04/2019
 # Author: Callum Bruce
 # Suborbital Falcon9 example
-
 import numpy as np
-import matplotlib.pyplot as plt
-from time import time
-from main import ReferenceFrame, Body, Stage, Vehicle
-from forcetorque import gravity, thrust
-from simulate import simulate, euler
-from control import PIDcontroller
-
-startTime = time()
+from mayavi import mlab
+from pyquaternion import Quaternion
+from rocketsim import *
 
 # Define Earth
-earthRF = ReferenceFrame()
-earth = Body(5.972e24,6.371e6,[0,0,0,0,0,0],earthRF)
+earth = CelestialBody(5.972e24, 6.371e6)
 
 # Define Falcon9
-falcon9RF = ReferenceFrame()
-stage1 = Stage(258500,1.85,13.1,[-6.55,0])
-stage2 = Stage(52000,1.85,35,[-30.6,0])
-falcon9 = Vehicle([stage1,stage2],[0,0,earth.radius+48.1,0,0,0],earthRF,falcon9RF)
+stage1 = Stage(258500, 1.85, 35, [-30.6, 0, 0])
+stage2 = Stage(52000, 1.85, 13.1, [-6.55, 0, 0])
+falcon9 = Vessel([stage1, stage2], parent=earth)
+falcon9.setPosition([earth.radius + 48.1, 0, 0])
+falcon9.updateNorthEastDownRF()
+falcon9.initAttitude()
+# Plotting
+figure = mlab.figure(size=(600, 600))
+
+earthImageFile = 'plotting/earth.jpg'
+#plotCelestialBody(figure, earth.getRadius(), earth.getPosition(), earthImageFile)
+
+#falcon9.updateNorthEastDownRF()
+falcon9.northeastdownRF.plot(figure, falcon9.getPosition(), scale_factor=100000)
+
+### Setup Initial bodyRF position ###
+#falcon9.initAttitude()
+#falcon9.bodyRF.plot(figure, falcon9.getPosition(), scale_factor=200000)
+
+falcon9.addTorque([531343.125 * np.deg2rad(0), 6.01303303e+07 * np.deg2rad(0), 6.01303303e+07 * np.deg2rad(175)])
+#falcon9.addTorque([531343.125 * np.deg2rad(95), 6.01303303e+07 * np.deg2rad(160), 6.01303303e+07 * np.deg2rad(-50)])
+#falcon9.addForce([310500 * 100, 0, 0], local=True)
+simulate(falcon9, euler, 1)
+dt = 0.001
+for i in range(0,1000):
+    simulate(falcon9, euler, dt)
+falcon9.bodyRF.plot(figure, falcon9.getPosition(), scale_factor=200000)
+
+mlab.view(focalpoint=falcon9.getPosition(), figure=figure)
+
+figure1 = mlab.figure(size=(600, 600))
+universalAttitude = falcon9.getAttitude()
+universalAttitudeRF = ReferenceFrame()
+universalAttitudeRF.plot(figure1, [0, 0, 0])
+universalAttitudeRF_new = ReferenceFrame()
+universalAttitudeRF_new.rotate(universalAttitude)
+universalAttitudeRF_new.plot(figure1, [0, 0, 0], scale_factor=2)
+
+figure2 = mlab.figure(size=(600, 600))
+localAttitude = falcon9.getAttitude(local=True)
+localAttitudeRF = ReferenceFrame()
+localAttitudeRF.plot(figure2, [0, 0, 0])
+localAttitudeRF_new = ReferenceFrame()
+localAttitudeRF_new.rotate(localAttitude)
+localAttitudeRF_new.plot(figure2, [0, 0, 0], scale_factor=2)
+
+### Gimbal Control: ###
+# Step 1: Work out the torque vector required to turn onto a heading.
+# Step 2: Find unit_res (at CoT) in bodyRF equivelant to apply torque in direction of torque vector.
+# Step 3: Find gimbal angle to achieve unit_res.
+# Step 4: Based on Kp, Ki and Kd turn onto heading.
+'''
 Isp = 300 # (s)
 m_dot = 1500 # (kg/s)
 
 # Define controller
-gains = [0.12,0.01,0.35]
-lims = [5,-5]
+gains = [0.12, 0.01, 0.35]
+lims = [5, -5]
 windup = 0.2
-yawControl = PIDcontroller(gains,lims,windup)
+yawControl = PIDcontroller(gains, lims, windup)
 d = np.array([0.0]) # Initial gimbal angle (deg)
 
 # Simulation loop
@@ -46,7 +87,7 @@ for i in range(0,8000):
 
     # Simulate falcon9
     simulate(falcon9,falcon9RF,euler,dt)
-    '''
+
     # Fix attitude
     if i > 300 and i < 900:
         att = np.deg2rad(20)
@@ -56,7 +97,7 @@ for i in range(0,8000):
         att = np.arctan(v[1]/v[0])
         falcon9.setPhi(att)
         print(np.rad2deg(att))
-    '''
+
     # Use PID to control phi
     if i > 300 and i < 900: # Control phi = 20 deg
         phi_pv = falcon9.getPhi()
@@ -77,52 +118,4 @@ for i in range(0,8000):
     forceThrust, torqueThrust = thrust(earth,falcon9,m_dot,Isp,d[-1],dt) # forceThrust, torqueThrust from fuel burn
     forceThrustSave = np.append(forceThrustSave,[forceThrust],axis=0)
     falcon9.appendU(list(np.append(forceGravity+forceThrust,torqueThrust)))
-
-# Plotting
-state = np.array(falcon9.state)
-falcon9Pos = falcon9.getPosition()
-fig, ax = plt.subplots()
-ax.axis('equal')
-ax.axis([-earth.radius*2,earth.radius*2,-earth.radius*2,earth.radius*2])
-earthPosition = earth.getPosition()
-earthPlot = plt.Circle((earthPosition[0], earthPosition[1]),earth.radius,color='b')
-ax.add_artist(earthPlot)
-ax.plot(state[0:,2],state[0:,3],color='k',lw=0.5)
-ax.plot([falcon9Pos[0],falcon9Pos[0]+falcon9RF.i[0]*1000000],[falcon9Pos[1],falcon9Pos[1]+falcon9RF.i[1]*1000000],c='r')
-ax.plot([falcon9Pos[0],falcon9Pos[0]+falcon9RF.j[0]*1000000],[falcon9Pos[1],falcon9Pos[1]+falcon9RF.j[1]*1000000],c='g')
-
-# Plot state over time [d, phi, phi_dot]
-fig, ax1 = plt.subplots(3,1)
-ax1[0].plot(t,d)
-ax1[0].set_title('d')
-ax1[0].set_ylabel('[rad]')
-ax1[0].grid()
-ax1[1].plot(t,state[0:,5])
-ax1[1].set_title('phi')
-ax1[1].set_ylabel('[rad]')
-ax1[1].grid()
-ax1[2].plot(t,state[0:,4])
-ax1[2].set_title('phi_dot')
-ax1[2].set_ylabel('[rad/s]')
-ax1[2].grid()
-
-# Plot controller [phi_SP phi_PV, error, int_error]
-fig, ax2 = plt.subplots(3,1)
-ax2[0].plot(yawControl.SP)
-ax2[0].plot(yawControl.PV)
-ax2[0].set_title('phi')
-ax2[0].set_ylabel('[rad]')
-ax2[0].legend(['Set Point','State'])
-ax2[0].grid()
-ax2[1].plot(yawControl.error)
-ax2[1].set_title('error')
-ax2[1].set_ylabel('[rad]')
-ax2[1].grid()
-ax2[2].plot(yawControl.int)
-ax2[2].set_title('int_error')
-ax2[2].set_ylabel('[rad]')
-ax2[2].grid()
-plt.show()
-
-endTime = time()
-print(str(endTime - startTime) + ' Seconds')
+'''
