+++ rocketsim/plotting/earth.jpg
