+++ rocketsim/plotting/navball_brownblue.jpg
