+++ rocketsim/plotting/plotting.py
@@ -1,8 +1,12 @@
+# Date: 02/04/2019
+# Author: Callum Bruce
+# Plotting functions
 from mayavi import mlab
 from tvtk.api import tvtk # python wrappers for the C++ vtk ecosystem
 import numpy as np
 ## See www.cgtrader.com for textures ###
-def plotBody(figure, radius, position, image):
+
+def plotCelestialBody(figure, radius, position, image):
     """
     Plot body and map texture to sphere using mayavi and tvtk.
 
@@ -12,17 +16,17 @@ def plotBody(figure, radius, position, image):
         position (list): Body position [x, y, z].
         image (string): String specifying .jpeg for body texture.
     """
-    # Load and map texture
+    # Step 1: Load and map texture
     img = tvtk.JPEGReader()
     img.file_name = image
     texture = tvtk.Texture(input_connection=img.output_port, interpolate=1)
-    # Create a TexturedSphereSource with a given radius and angular resolution
+    # Step 2: Create a TexturedSphereSource with a given radius and angular resolution
     Nrad = 180
     sphere = tvtk.TexturedSphereSource(radius=radius, theta_resolution=Nrad, phi_resolution=Nrad) # Pipeline - source
     sphere_mapper = tvtk.PolyDataMapper(input_connection=sphere.output_port) # Pipeline - mapper
     sphere_actor = tvtk.Actor(mapper=sphere_mapper, texture=texture) # Pipeline - actor
     sphere_actor.add_position((position[0],position[1],position[2])) # Pipeline - actor.add_position
-    # Add actor to figure
+    # Step 3: Add actor to figure
     figure.scene.add_actor(sphere_actor)
 
 def plotTrajectory(figure, points, color):
@@ -40,23 +44,17 @@ def plotTrajectory(figure, points, color):
     line_actor = tvtk.Actor(mapper=line_mapper, property=p)
     figure.scene.add_actor(line_actor)
 
-# Define figure
-figure = mlab.figure(size=(600, 600))
+def plotCylinder(figure, radius, height, center):
+    """
+    Plot cylinder using mayavi and tvtk.
 
-# Plot earth
-earthImageFile = 'earth.jpg'
-earthPosition = [0, 0, 0]
-earthRadius = 6.371e6
-plotBody(figure, earthRadius, earthPosition, earthImageFile)
-
-# Plot trajectory
-points = np.load('iss_points.npy')
-color = (1, 1, 1)
-plotTrajectory(figure, points, color)
-
-'''
-Note: Can animate rotating Earth by looping...
-sphere_actor.add_orientation((0,0,5))
-fig.scene.add_actor(sphere_actor)
-...every n seconds
-'''
+    Args:
+        figure (mlab.figure): Mayavi figure for plot.
+        radius (float): Cylinder radius (m).
+        height (float): Cylinder height (m).
+        center (list): Cylinder center [x, y, z] (m).
+    """
+    cylinder = tvtk.CylinderSource(radius=radius, height=height, center=[-center[1], center[0], center[2]], resolution=90)
+    cylinder_mapper = tvtk.PolyDataMapper(input_connection=cylinder.output_port)
+    cylinder_actor = tvtk.Actor(mapper=cylinder_mapper, orientation=[0, 0, -90])
+    figure.scene.add_actor(cylinder_actor)
